+++ itolib/Behaviours/Props/GiftSpawner.cs
@@ -118,8 +118,9 @@ namespace itolib.Behaviours.Props
             int minValue = giftContentsMinValue < 0 ? gift.objectInPresentItem.minValue : giftContentsMinValue,
                 maxValue = giftContentsMaxValue < 0 ? gift.objectInPresentItem.maxValue : giftContentsMaxValue;
 
-            int scrapValue = isSeededRandom ? SeededSelf.GetSeededRandom().Next(minValue, maxValue)
-                : Random.RandomRangeInt(minValue, maxValue);
+            int scrapValue = (minValue >= maxValue) ? minValue
+                : (isSeededRandom ? SeededSelf.GetSeededRandom().Next(minValue, maxValue)
+                : Random.RandomRangeInt(minValue, maxValue));
 
             if (RoundManager.Instance != null)
             {